GH-4 Update reporter to use ICucumberFeatureSuite model
